refactoring: replace exception classes
