+++ order-microservice/src/main/java/com/wasd/ordermicroservice/exception/common/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.wasd.ordermicroservice.exception;
+package com.wasd.ordermicroservice.exception.common;
 
 public class NotFoundException extends Exception {
