+++ order-microservice/src/main/java/com/wasd/ordermicroservice/exception/persistence/OrderCreationException.java
@@ -1,4 +1,4 @@
-package com.wasd.ordermicroservice.exception;
+package com.wasd.ordermicroservice.exception.persistence;
 
 public class OrderCreationException extends RuntimeException {
     public OrderCreationException() {
