+++ order-microservice/src/main/java/com/wasd/ordermicroservice/rest/SellerController.java
@@ -2,8 +2,8 @@ package com.wasd.ordermicroservice.rest;
 
 import com.wasd.ordermicroservice.data.seller.SellerRequest;
 import com.wasd.ordermicroservice.data.seller.SellerResponse;
-import com.wasd.ordermicroservice.exception.AlreadyExistsException;
-import com.wasd.ordermicroservice.exception.NotFoundException;
+import com.wasd.ordermicroservice.exception.common.AlreadyExistsException;
+import com.wasd.ordermicroservice.exception.common.NotFoundException;
 import com.wasd.ordermicroservice.service.seller.SellerService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
