+++ order-microservice/src/main/java/com/wasd/ordermicroservice/service/brand/BrandService.java
@@ -2,7 +2,7 @@ package com.wasd.ordermicroservice.service.brand;
 
 import com.wasd.ordermicroservice.data.brand.BrandRequest;
 import com.wasd.ordermicroservice.data.brand.BrandResponse;
-import com.wasd.ordermicroservice.exception.NotFoundException;
+import com.wasd.ordermicroservice.exception.common.NotFoundException;
 
 public interface BrandService {
     BrandResponse create(BrandRequest request);
