+++ order-microservice/src/main/java/com/wasd/ordermicroservice/service/order/OrderServiceImpl.java
@@ -2,8 +2,8 @@ package com.wasd.ordermicroservice.service.order;
 
 import com.wasd.ordermicroservice.data.order.OrderRequest;
 import com.wasd.ordermicroservice.data.order.OrderResponse;
-import com.wasd.ordermicroservice.exception.NotFoundException;
-import com.wasd.ordermicroservice.exception.OrderCreationException;
+import com.wasd.ordermicroservice.exception.common.NotFoundException;
+import com.wasd.ordermicroservice.exception.persistence.OrderCreationException;
 import com.wasd.ordermicroservice.persistence.order.Order;
 import com.wasd.ordermicroservice.persistence.order.OrderRepository;
 import com.wasd.ordermicroservice.util.mapper.OrderMapper;
