+++ order-microservice/src/main/java/com/wasd/ordermicroservice/service/seller/SellerService.java
@@ -2,8 +2,8 @@ package com.wasd.ordermicroservice.service.seller;
 
 import com.wasd.ordermicroservice.data.seller.SellerRequest;
 import com.wasd.ordermicroservice.data.seller.SellerResponse;
-import com.wasd.ordermicroservice.exception.AlreadyExistsException;
-import com.wasd.ordermicroservice.exception.NotFoundException;
+import com.wasd.ordermicroservice.exception.common.AlreadyExistsException;
+import com.wasd.ordermicroservice.exception.common.NotFoundException;
 
 public interface SellerService {
     SellerResponse findById(Long id) throws NotFoundException;
